+++ services/QuillLMS/client/app/bundles/Teacher/components/accounts/new/__tests__/verify_school.test.tsx
@@ -5,8 +5,14 @@ import VerifySchool from '../verify_school';
 
 describe('VerifySchool component', () => {
 
-  it('should render', () => {
-    const component = mount(<VerifySchool passedSchoolName="School Name" />);
+  it('should render if there are no admins', () => {
+    const component = mount(<VerifySchool passedAdmins={[]} passedSchoolName="School Name" />);
     expect(component).toMatchSnapshot();
   });
+
+  it('should render if there are admins', () => {
+    const component = mount(<VerifySchool passedAdmins={[{ name: 'Joy Harjo', email: 'joyharjo@quill.org', id: 1 }]} passedSchoolName="School Name" />);
+    expect(component).toMatchSnapshot();
+  });
+
 });
