+++ services/QuillLMS/client/app/bundles/Teacher/components/accounts/new/skip_for_now.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react'
+
+const SkipForNow = () => (<a className="skip-for-now" href="/finish_sign_up">Skip for now</a>)
+
+export default SkipForNow
